# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tabanni Pet Adoption Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gallery:
    get:
      operationId: listGalleryPosts
      tags: [gallery]
      summary: List gallery posts
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of gallery posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPost"
    post:
      operationId: createGalleryPost
      tags: [gallery]
      summary: Create a gallery post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryPostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPost"
  /gallery/{id}:
    get:
      operationId: getGalleryPost
      tags: [gallery]
      summary: Get a single gallery post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Gallery post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPost"
...
components:
  schemas:
...
    GalleryPost:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        authorId:
          type: number
          nullable: true
        authorName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - createdAt
    CreateGalleryPostInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: string
        authorId:
          type: number
      required:
        - title
        - content
...